Moving to a slide index which is not a multiple of slidesToScroll.
ACHIEVEMENTS

Sliders might be navigated by dragging, swiping, pressing the arrow
buttons and, last but not least, using a custom slide category
navigator.

Let's imagine our slides to be divided in different "groups" (let's name
them "categories"). Let's also imagine a list of clickable items (the
category list, actually) in the nearabouts of our slider.

We reasonably want to achieve the following behaviours:
- If the slider is explored bw swiping/dragging/pressing the arrow
buttons, I want it to scroll according to the slidesToScroll parameters;
- If I click on a category, I want the slider to go EXACTLY to the first
slide of the category. Currently (version 1.6.0), the slider moves to
the previous "navigable item", i.e. the slide whose index is the
greatest index which is a multiple of slidesToScroll and which is lesser
than the desired position);
- If I click on a category, then I swipe/drag/use the arrow buttons
again, I want the slider to scroll according to the pre-computed
"navigable items" again.


SOLUTION SUMMARY

We added a parameter named "ignoreSlidesToScroll", to be set to "true"
and passed to the slider when using "goTo", "slickPrev" and "slickNext"
methods. If set to true, the slider will move exactly to the specified
slider index, actually ignoring the slidesToScroll value and the
pre-computed navigable items. The new parameter is handled in a
completely backward-compatible way.


EXAMPLE

A slider contains 17 photos, divided into two categories: cat1 and cat2.
Cat1 starts at index 0, cat2 starts at index 7.
I want to show 5 slides and scroll by 5 slides.

Current behaviour:
When I click on cat2 (index 7), the slider actually moves to the index
5.

Modified behaviour:
When I click on cat2 (index 7), the slider actually moves to the index
7.
If I subsequently click on the back button, the slider moves to the
index 5. Otherwise, if I click on the next button, the slider moves to
the index 10.